requirements.txt
numpy
torch>=1.0
torchvision
matplotlib
tensorboardx
pillow
visdom
opencv
